+++ src/managers/refresh-ahead.ts
@@ -45,7 +45,15 @@ class RefreshAheadManager extends BaseManager {
   public async get<E extends Executor>(key: string, executor: E, options: ReadWriteOptions = {}):
     Promise<ValueOfExecutor<E>> {
     const executorContext = { key, executor, options };
-    const record = await this.storage.get(key);
+    let record: StorageRecord | null = null;
+
+    try {
+      record = await this.storage.get(key);
+    } catch (e) {
+      this.logger.error('Failed to get value from storage, falling back to executor', e);
+
+      return executor();
+    }
 
     if (await this.isRecordValid(record)) {
       const result = deserialize((record as any).value);