+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+#### [1.1.1]
+- run executor after storage throws an error
+- change default operation timeout to 150
+
 #### [1.1.0]
 - support for dynamic tags in Manager's default storage via "getTags" option.
 - updated jest